reorganize base components
