+++ src/components/Projects.tsx
@@ -1,5 +1,5 @@
 import { ProjectCardProps } from "@/types/data";
-import Typography from "./base/Typography";
+import Typography from "@/components/base/Typography/Typography";
 import ProjectCard from "./ProjectCard";
 
 const Projects = ({
