+++ src/components/base/Button/Button.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { VariantProps, cva } from "class-variance-authority";
-import { cn } from "./utils";
+import { cn } from "../utils";
 
 const buttonVariants = cva("p-2 rounded-lg cursor-pointer", {
   variants: {
