+++ src/components/base/Typography/Typography.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { VariantProps, cva } from "class-variance-authority";
-import { cn } from "./utils";
+import { cn } from "../utils";
 
 const typographyVariants = cva("", {
   variants: {
