+++ src/components/layout/ThemeButton.tsx
@@ -1,6 +1,6 @@
 import { FaRegMoon } from "react-icons/fa";
 import { FiSun } from "react-icons/fi";
-import Button from "../base/Button";
+import Button from "@/components/base/Button/Button";
 import { useTheme } from "@/hooks/useTheme";
 
 const ThemeButton = () => {
